Update modeldict query sets for Django >= 1.9 with backwards compatibility.
